mtce

+++ update_resources/connectors/wikimedia_comnames.php
@@ -1,6 +1,10 @@
 <?php
 namespace php_active_record;
-/* DATA-1770: separate wikimedia into an images resource and a common names resource */
+/* DATA-1770: separate wikimedia into an images resource and a common names resource 
+This will rename 71.tar.gz from the long Jenkins run to 71_orig.tar.gz
+The remaining 71.tar.gz is the resource with just images.
+The wikimedia_comnames.tar.gz is the resource with just vernaculars.
+*/
 
 // ini_set('error_reporting', E_ALL);
 // ini_set('display_errors', true);
@@ -11,22 +15,24 @@ $timestart = time_elapsed();
 require_library('connectors/DwCA_Utility');
 ini_set('memory_limit','7096M'); //required
 
-// $resource_id = 'wikimedia_comnames';
-// $func = new DwCA_Utility($resource_id, CONTENT_RESOURCE_LOCAL_PATH . "71" . ".tar.gz");
-// $preferred_rowtypes = array('http://rs.tdwg.org/dwc/terms/taxon', 'http://rs.gbif.org/terms/1.0/vernacularname');
-// $func->convert_archive($preferred_rowtypes);
-// Functions::finalize_dwca_resource($resource_id);
-// $func = '';
+$resource_id = 'wikimedia_comnames';
+$func = new DwCA_Utility($resource_id, CONTENT_RESOURCE_LOCAL_PATH . "71" . ".tar.gz");
+$preferred_rowtypes = array('http://rs.tdwg.org/dwc/terms/taxon', 'http://rs.gbif.org/terms/1.0/vernacularname');
+$func->convert_archive($preferred_rowtypes);
+Functions::finalize_dwca_resource($resource_id, false, true); //2nd param false means not a big file, 3rd param true means delete working folder in CONTENT_RESOURCE_LOCAL_PATH
+$func = '';
 
 $resource_id = '71_new';
 $func = new DwCA_Utility($resource_id, CONTENT_RESOURCE_LOCAL_PATH . "71" . ".tar.gz");
 $preferred_rowtypes = array('http://rs.tdwg.org/dwc/terms/taxon', 'http://eol.org/schema/agent/agent', 'http://eol.org/schema/media/document');
 $func->convert_archive($preferred_rowtypes);
-Functions::finalize_dwca_resource($resource_id);
+Functions::finalize_dwca_resource($resource_id, false, true);
 
-if(file_exists(CONTENT_RESOURCE_LOCAL_PATH . "71_orig.tar.gz")) unlink(CONTENT_RESOURCE_LOCAL_PATH . "71_orig.tar.gz");
-Functions::file_rename(CONTENT_RESOURCE_LOCAL_PATH . "71.tar.gz", CONTENT_RESOURCE_LOCAL_PATH . $resource_id . "71_orig.tar.gz");
-Functions::file_rename(CONTENT_RESOURCE_LOCAL_PATH . "71_new.tar.gz", CONTENT_RESOURCE_LOCAL_PATH . $resource_id . "71.tar.gz");
+if(file_exists(CONTENT_RESOURCE_LOCAL_PATH . "71_new.tar.gz")) {
+    if(file_exists(CONTENT_RESOURCE_LOCAL_PATH . "71_orig.tar.gz")) unlink(CONTENT_RESOURCE_LOCAL_PATH . "71_orig.tar.gz");
+    Functions::file_rename(CONTENT_RESOURCE_LOCAL_PATH . "71.tar.gz", CONTENT_RESOURCE_LOCAL_PATH . "71_orig.tar.gz");
+    Functions::file_rename(CONTENT_RESOURCE_LOCAL_PATH . "71_new.tar.gz", CONTENT_RESOURCE_LOCAL_PATH . "71.tar.gz");
+}
 
 $elapsed_time_sec = time_elapsed() - $timestart;
 echo "\n";
